requirements.txt
streamlit==1.28.2
paho-mqtt
pandas
